+++ src/main/java/org/wedding/domain/card/Card.java
@@ -1,8 +1,10 @@
-package org.wedding.domain;
+package org.wedding.domain.card;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+
+import org.wedding.domain.user.Todo;
 
 import lombok.Getter;
