+++ src/main/java/org/wedding/domain/user/Todo.java
@@ -1,4 +1,4 @@
-package org.wedding.domain;
+package org.wedding.domain.user;
 
 import lombok.Getter;
