+++ src/main/java/org/wedding/user/repository/mybatis/UserMapper.java
@@ -3,7 +3,7 @@ package org.wedding.user.repository.mybatis;
 import java.util.Optional;
 
 import org.apache.ibatis.annotations.Mapper;
-import org.wedding.user.domain.User;
+import org.wedding.domain.user.User;
 import org.wedding.user.repository.UserRepository;
 
 @Mapper
