+++ src/test/java/org/wedding/WeddingApplicationTests.java
@@ -1,6 +1,5 @@
-package project.wedding;
+package org.wedding;
 
-import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
