+++ src/test/java/org/wedding/domain/CardTest.java
@@ -1,11 +1,11 @@
-package project.wedding.domain;
+package org.wedding.domain;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.wedding.domain.Card;
+import org.wedding.domain.card.Card;
 
 @Disabled
 class CardTest {
