+++ src/test/java/org/wedding/user/controller/UserControllerTest.java
@@ -1,9 +1,10 @@
-package project.wedding.user.controller;
+package org.wedding.user.controller;
 
 import static org.mockito.Mockito.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
+import static org.wedding.user.exception.UserError.*;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
